--- src/main/java/project/config/implementation/storage/PrimitiveStorage.java
@@ -1,6 +1,0 @@
-package project.config.implementation.storage;
-
-import project.config.api.Storage;
-
-public abstract class PrimitiveStorage extends Storage {
-}
