+++ src/main/java/project/save/api/Serializable.java
@@ -1,4 +1,4 @@
-package project.config.api;
+package project.save.api;
 
 import java.util.HashMap;
