+++ src/main/java/project/save/api/Storage.java
@@ -1,4 +1,4 @@
-package project.config.api;
+package project.save.api;
 
 public abstract class Storage {
     public abstract Object read();
