+++ src/main/java/project/save/implementation/storage/PrimitiveStorage.java
@@ -1,0 +1,6 @@
+package project.save.implementation.storage;
+
+import project.save.api.Storage;
+
+public abstract class PrimitiveStorage extends Storage {
+}
